Evict the PIT.
